+++ src/api-sdk/contracts/future-api-sdk-evolution-profile.js
@@ -1,0 +1,27 @@
+export class FutureApiSdkEvolutionProfile {
+  constructor({ profileName, objectives = [], strategicDrivers = [], invariants = [], horizons = [], maturityLevels = [],
+    capabilityAreas = [], experimentStates = [], adoptionGates = [], evidenceTypes = [], stopCriteria = [],
+    architecturalRules = [], architectureBoundaries = [], invariantsPreserved = true, businessFirst = true,
+    authoritativeProviderContracts = true, discoveryNoAuthority = true, tenantIsolationPreserved = true,
+    generatedArtifactsTraceable = true, humanAccountability = true, evidenceBasedPromotion = true,
+    experimentsControlled = true, compatibilityMigrationGoverned = true, operationsAndRetirementRequired = true,
+    boundedAutomation = true, portabilityExitTested = true, technologyNeutral = true, adrDecisions = true,
+    technologyFirst = false, discoveryGrantsAuthority = false, tenantForksDefault = false, permanentPreview = false,
+    aiAutonomousApproval = false, universalExactlyOnce = false, marketplaceListingProvesTrust = false,
+    experimentProductionApproved = false, portabilityClaimUntested = false, selectsTechnology = false,
+    replacesChapterAuthorities = false }) {
+    this.profileName = profileName;
+    for (const [key, value] of Object.entries({ objectives, strategicDrivers, invariants, horizons, maturityLevels,
+      capabilityAreas, experimentStates, adoptionGates, evidenceTypes, stopCriteria, architecturalRules, architectureBoundaries })) {
+      this[key] = Object.freeze([...value]);
+    }
+    for (const [key, value] of Object.entries({ invariantsPreserved, businessFirst, authoritativeProviderContracts,
+      discoveryNoAuthority, tenantIsolationPreserved, generatedArtifactsTraceable, humanAccountability,
+      evidenceBasedPromotion, experimentsControlled, compatibilityMigrationGoverned, operationsAndRetirementRequired,
+      boundedAutomation, portabilityExitTested, technologyNeutral, adrDecisions, technologyFirst,
+      discoveryGrantsAuthority, tenantForksDefault, permanentPreview, aiAutonomousApproval, universalExactlyOnce,
+      marketplaceListingProvesTrust, experimentProductionApproved, portabilityClaimUntested, selectsTechnology,
+      replacesChapterAuthorities })) this[key] = Boolean(value);
+    Object.freeze(this);
+  }
+}